Extract result_modifiers code from obfuscation
