+++ src/hutch_bunny/core/obfuscation.py
@@ -1,26 +1,4 @@
-import json
 from typing import Union
-
-
-def get_results_modifiers_from_str(params: str) -> list:
-    """Deserialise a JSON list containing results modifiers
-
-    Args:
-        params (str):
-        The JSON string containing list of parameter objects for results modifiers
-
-    Raises:
-        ValueError: The parsed string does not produce a list
-
-    Returns:
-        list: The list of parameter dicts of results modifiers
-    """
-    deserialised_params = json.loads(params)
-    if not isinstance(deserialised_params, list):
-        raise ValueError(
-            f"{get_results_modifiers_from_str.__name__} requires a JSON list"
-        )
-    return deserialised_params
 
 
 def low_number_suppression(
