+++ src/hutch_bunny/core/results_modifiers.py
@@ -1,3 +1,6 @@
+import json
+
+
 def results_modifiers(
     low_number_suppression_threshold: int,
     rounding_target: int,
@@ -18,3 +21,24 @@ def results_modifiers(
             }
         )
     return results_modifiers
+
+
+def get_results_modifiers_from_str(params: str) -> list:
+    """Deserialise a JSON list containing results modifiers
+
+    Args:
+        params (str):
+        The JSON string containing list of parameter objects for results modifiers
+
+    Raises:
+        ValueError: The parsed string does not produce a list
+
+    Returns:
+        list: The list of parameter dicts of results modifiers
+    """
+    deserialised_params = json.loads(params)
+    if not isinstance(deserialised_params, list):
+        raise ValueError(
+            f"{get_results_modifiers_from_str.__name__} requires a JSON list"
+        )
+    return deserialised_params
